scrapper/requirements.txt
redis
requests
sentry-sdk
python-dotenv
schedule